change session storage format
We only had one slot in the session for storing info, meaning that trying to deal with multiple surveys simultaneously caused them to overwrite each other.  Use a UUID-based partitioning to solve this.

WARNING: This is a breaking change in the way that we store session information, and will break sessions in progress.  Try to deploy this during a time of low usage.